requirements.txt
flask
jsonschema
azure-storage-blob
azure-identity
openai
pandas
scikit-learn
shap
python-docx
openpyxl
sentence-transformers
fastapi
uvicorn
joblib
requests
azure-ai-agents
azure-ai-projects
gunicorn
asyncio
semantic-kernel
docx2txt
ngrok
fastmcp